requirements.txt
fastapi==0.103.1
keras==2.15.0
numpy==1.26.4
protobuf==3.20.3
pydantic==2.9.2
uvicorn==0.22.0
python-multipart==0.0.6
pillow==10.4.0
tensorflow-cpu==2.15.0
streamlit

